Bug 2068982 - Move the autofill bulk-import helpers to db/mod.rs
with_savepoint, timestamp_from_millis and CounterUpdate were added to
addresses.rs for the address with-meta import. Credit cards need all three for
the same reason, and none of them is address-specific, so share them rather
than duplicating.

No behaviour change: with_savepoint goes from module-private to pub(crate) and
CounterUpdate's re-export moves from crate::db::addresses to crate::db, which is
the only caller change.
